% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{base64EncFile}
\alias{base64EncFile}
\title{Method for base64 encoding}
\usage{
base64EncFile(file_name)
}
\arguments{
\item{fileName}{String. Name of the file to be encoded.}
}
\value{
Base 64 eencoded file.
}
\description{
Helper method for base64 encoding.
 Needed to transfer network and dataset files-
}
