% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful_utils.R
\name{call.KPM}
\alias{call.KPM}
\title{Method used to create a job submission}
\usage{
call.KPM(indicator_matrices, attached_to_id = NULL, async = TRUE, ...)
}
\arguments{
\item{indicator_matrices}{List. Indicator matrices that will be used.}

\item{attached_to_id}{String. The quest id this job has been attached to.}

\item{async}{Boolean. Submit a new job asynchronously?}

\item{...}{Options. Run parameters.}
}
\value{
Results object with runId, resultGraphs, comment, succes,resultUrl
 and questId
}
\description{
Method used to create a job submission
}
\examples{
call.KPM(list(huntington_disease_up), attached_to_id,
        async=FALSE, Lmin=8, Kmin=1, strategy="INES",
        removeBENs=TRUE, graphID=I2D.id, graph.file=NULL,
        range=FALSE, linkType="OR")
}
