% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_utils.R
\name{call_kpm_local}
\alias{call_kpm_local}
\title{Method for running Rkpm localy}
\usage{
call_kpm_local(indicator_matrices, graph_file)
}
\arguments{
\item{indicator_matrices}{List of paths to the indicator matrices that will be used.}

\item{graph_file}{String. Path to graph_file.}
}
\value{
Saves the results in folder results.
}
\description{
Method for running Rkpm localy
}
