% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpm.R
\name{check_files}
\alias{check_files}
\title{Function which checks and processes files}
\usage{
check_files(indicator_matrices, graph)
}
\arguments{
\item{indicator_matrices}{Filepath, data.frame or as a list of both.}

\item{graph}{Filepath to the graph file in sif format or an igraph object.}
}
\value{
Returns the data prepared for the respective run (local or remote).
Indicator matrices: List of paths for local run and list of data.frames for remote run.
Graph file: Null if no graph_file provided else path to graph_file.
}
\description{
Goes through all matrices and the graph file and checks
if the input data is correct and exists. It then prepares
the data for local or remote use. For local execution return
filepaths for remote reutrn data.frames. For the graph_files
always return filepaths.
}
