% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpm.R
\name{check_parameters}
\alias{check_parameters}
\title{Function which check parameters}
\usage{
check_parameters(indicator_matrices)
}
\arguments{
\item{indicator_matrices}{List of paths to the indicator matrices.}
}
\description{
Checks case and gene exception parameters
as well as perturbation paramets. Checks sanity
of parameters and if they are provided in the correct
form.
}
