% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpm.R
\name{combine_mutliple_matrices}
\alias{combine_mutliple_matrices}
\title{Logically combines multiple datasets}
\usage{
combine_mutliple_matrices(indicator_matrices)
}
\arguments{
\item{indicator_matrices}{Indicator matrices to combine}
}
\value{
Indicator matrix
}
\description{
Logically combines multiple datasets
}
