% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_z_scores}
\alias{compute_z_scores}
\title{Computes z-score values of case samples}
\usage{
compute_z_scores(norm_counts, controls, cases)
}
\arguments{
\item{norm_counts}{Normalized counts as data.frame in TMM, TPM or FPKM.}

\item{controls}{Vector with the column numbers of the control samples.}

\item{cases}{Vector with the column numbers of the case samples.}
}
\value{
Data.frame with the z.score values of the case samples.
}
\description{
Computes z-scores of the genes in the case samples
in contrast to the control samples.
}
\examples{
compute_z_scores(norm_counts, c(1, 2), c(3, 4))
}
