% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{Configuration-class}
\alias{Configuration-class}
\title{A configuration of the parameters K and L}
\description{
Contains the configuration and all the
correspondig graphs.
}
\section{Slots}{

\describe{
\item{\code{configuration}}{Name of the configuration (e.g. K-5-L1-15).}

\item{\code{k}}{Number of node exceptions.}

\item{\code{l_values}}{Number of case exceptions. Each dataset corresponds to a case_exception parameter.}
}}

