% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{create_indicator_from_fcurdleTest}
\alias{create_indicator_from_fcurdleTest}
\title{Creates indicator matrix from differentail}
\usage{
create_indicator_from_fcurdleTest(
  fcHurdletest_filtered,
  sc_obj,
  nameIndicatorMatrix = NULL
)
}
\arguments{
\item{fcHurdletest_filtered}{filtered differential expression analysis result}

\item{sc_obj}{single cell object used for differential expression analysis}

\item{nameIndicatorMatrix}{name of indicator matrix to be given}
}
\value{
indicator matrix in form of a data.frame
}
\description{
Creates indicator matrix from differentail
}
