% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{create_sca.Seurat}
\alias{create_sca.Seurat}
\title{Converts Seurat object to SingleCellAssay object}
\usage{
\method{create_sca}{Seurat}(seurat_object)
}
\arguments{
\item{seurat_object}{}
}
\value{
SingleCellAssay object
}
\description{
Raw counts from the Seurat object are log transformed
and used for the creation of a SingleCellAssay object
}
