% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{create_sca.SingleCellExperiment}
\alias{create_sca.SingleCellExperiment}
\title{Converts SingleCellExperiment object to SingleCellAssay object}
\usage{
\method{create_sca}{SingleCellExperiment}(sc_object)
}
\arguments{
\item{sc_object}{SingleCellExperiment object}
}
\value{
SingleCellAssay object
}
\description{
Converts SingleCellExperiment object to SingleCellAssay object
}
