% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{create_union_network}
\alias{create_union_network}
\title{Creates a union network for the pathways of a specific configuration}
\usage{
create_union_network(configuration)
}
\arguments{
\item{configuration}{Configuration object for which the union network should be computed.}
}
\value{
The union network as a pathway object
}
\description{
Creates a union network for the pathways of a specific configuration
}
