% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{dataset_list}
\alias{dataset_list}
\title{Encodes a list of datasets}
\usage{
dataset_list(indicator_matrices, attached_to_id)
}
\arguments{
\item{indicator_matrices}{List of indicator matrices to be encoded}

\item{attached_to_id}{Quest id for the current run.}
}
\value{
Encoded list of datasets with their respective questId.
}
\description{
Helper method to encode a list of datasets (indicator matrices)
 for the job submission.
}
