% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{determine_union_groups}
\alias{determine_union_groups}
\title{Determines from which pathways a node originates}
\usage{
determine_union_groups(union_nodes, configuration)
}
\arguments{
\item{configuration}{Configuration object for which the groups should be determined,}
}
\value{
Data frame with the node ids and the pathways they originate from
}
\description{
Determines from which pathways a node originates
}
