% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{do_diff_testing_via_MAST}
\alias{do_diff_testing_via_MAST}
\title{Calculates differentially expressed genes from single cell object}
\usage{
do_diff_testing_via_MAST(
  sc_obj,
  covarariate,
  referenceGroup,
  saveSummary = TRUE,
  summaryName = NULL,
  designFormula = NULL
)
}
\arguments{
\item{sc_obj}{single cell object of type seurat, singleCellAssay or singlCellExperiment}

\item{covarariate}{colname of metadata with groups for differential expression}

\item{referenceGroup}{reference/control group}

\item{saveSummary}{if TRUE saves the result of the differential analysis as a .rds file}

\item{summaryName}{name of the rds file if saveSummary=TRUE}

\item{designFormula}{string which will be used as formula in MAST's zlm function}
}
\value{
data.frame of differential expression analysis results
}
\description{
Calculates differentially expressed genes from single cell object
}
