% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix.R
\name{export_matrix}
\alias{export_matrix}
\title{Saves indicator matrix as a file.}
\usage{
export_matrix(matrix, path)
}
\arguments{
\item{matrix}{Indicator or numerical matrix.}

\item{path}{Path where the file should be saved.}
}
\description{
Saves indicator matrix as a file.
}
