% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{export_nodes}
\alias{export_nodes}
\alias{export_nodes,Pathway-method}
\title{Export node ids}
\usage{
export_nodes(pathway_object, file)

\S4method{export_nodes}{Pathway}(pathway_object, file)
}
\arguments{
\item{pathway_object}{pathway object of a configuration.}

\item{file}{path where the file is saved}
}
\description{
Export node ids
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Use write.table to export a graph.
}}

