% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{export_to_iGraph}
\alias{export_to_iGraph}
\alias{export_to_iGraph,Pathway-method}
\title{Get an iGraph object for a specific pathway}
\usage{
export_to_iGraph(pathway_object)

\S4method{export_to_iGraph}{Pathway}(pathway_object)
}
\arguments{
\item{pathway_object}{pathway object of a configuration.}
}
\value{
i_graph object of the pathway.
}
\description{
Get an iGraph object for a specific pathway
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Exports pathway object.
}}

