% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{fetch_results}
\alias{fetch_results}
\title{Once the run is complete, we can obtain the results}
\usage{
fetch_results(
  quest_id,
  url = "https://keypathwayminer.compbio.sdu.dk/keypathwayminer/"
)
}
\arguments{
\item{quest_id}{Jobs respective quest_id.}
}
\value{
If run was successful return result in json
format otherwise null.
}
\description{
Once the run is complete, we can obtain the results
}
