% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{filter_fcHurdleTestResults}
\alias{filter_fcHurdleTestResults}
\title{Filters the differential expression analysis results based on fold change and adjusted pvalue}
\usage{
filter_fcHurdleTestResults(
  fcHurdletest,
  FCThreshold = 2,
  pvalueThreshold = 0.05
)
}
\arguments{
\item{fcHurdletest}{result from do_diff_testing_via_MAST() function}

\item{FCThreshold}{threshold for the log 2 fold change}

\item{pvalueThreshold}{threshold for the adjusted p value}
}
\value{
data.frame of filtered differential expression analysis result
}
\description{
Filters the differential expression analysis results based on fold change and adjusted pvalue
}
