% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_utils.R
\name{get_case_exceptions}
\alias{get_case_exceptions}
\title{Crates java arguments for case exceptions}
\usage{
get_case_exceptions(indicator_matrices, arguments)
}
\description{
Crates java arguments for case exceptions
}
