% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_configuration}
\alias{get_configuration}
\alias{get_configuration,Result-method}
\title{Get configuration object}
\usage{
get_configuration(result_object, configuration_name)

\S4method{get_configuration}{Result}(result_object, configuration_name)
}
\arguments{
\item{result_object}{Result object obtained from a KeyPathwayMineR execution.}

\item{configuration_name}{Name of the configuration object.}
}
\value{
Configuration object.
}
\description{
Get configuration object
}
\section{Methods (by class)}{
\itemize{
\item \code{Result}: Given a result object and configuration name a configuration object is returned.
}}

