% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_configurations}
\alias{get_configurations}
\alias{get_configurations,Result-method}
\title{Get all configuration names of a result}
\usage{
get_configurations(result_object)

\S4method{get_configurations}{Result}(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
All configurations of a result.
}
\description{
Get all configuration names of a result
}
\section{Methods (by class)}{
\itemize{
\item \code{Result}: Retrun all configuration names of a results object.
}}

