% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{get_networks}
\alias{get_networks}
\title{Get a data frame of available networks}
\usage{
get_networks(url = "https://exbio.wzw.tum.de/keypathwayminer/")
}
\arguments{
\item{url}{Url to keypathwayminer.}
}
\value{
Data frame with different graphs and their Id's.
}
\description{
Get a data frame of available networks
}
\examples{
get_networks("https://exbio.wzw.tum.de/keypathwayminer/")
}
