% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_result_url}
\alias{get_result_url}
\alias{get_result_url,ResultRemote-method}
\title{Return result url}
\usage{
get_result_url(result_object)

\S4method{get_result_url}{ResultRemote}(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
Url to KPM web server.
}
\description{
Only for remote execution
}
\section{Methods (by class)}{
\itemize{
\item \code{ResultRemote}: Returns result URL
}}

