% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_results}
\alias{get_results}
\alias{get_results,ResultRemote-method}
\title{Fetches results once job is completed}
\usage{
get_results(result_object)

\S4method{get_results}{ResultRemote}(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
Remote result object.
}
\description{
Only for remote execution
}
\section{Methods (by class)}{
\itemize{
\item \code{ResultRemote}: returns RemoteResult object
}}

