% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_results_from_folder}
\alias{get_results_from_folder}
\title{Loads results from folder with text files}
\usage{
get_results_from_folder(path)
}
\arguments{
\item{path}{path to the folder with subfolders for each configuration}
}
\value{
KPM-R Result object
}
\description{
User can load result from "Result" folder which is automatically generated during each KPM run
}
