% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{get_status}
\alias{get_status}
\title{Method to check up on a submitted job.}
\usage{
get_status(quest_id)
}
\arguments{
\item{quest_id}{String. The quest id this job has been attached to.}
}
\value{
Status of the job for given quest_id.
}
\description{
Useful to monitor the job progress and current status.
}
