% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{getNetworks}
\alias{getNetworks}
\title{Get a data frame of available networks}
\usage{
getNetworks(url)
}
\arguments{
\item{url}{Url to keypathwayminer.}
}
\value{
Data frame with different graphs and their Id's.
}
\description{
Get a data frame of available networks
}
\examples{
getNetworks("https://exbio.wzw.tum.de/keypathwayminer/")
}
