% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{graph_kpm}
\alias{graph_kpm}
\title{Helper method to encode custom network file}
\usage{
graph_kpm(graph_file)
}
\arguments{
\item{graph_file}{Path to graph file which should be encoded.}
}
\value{
Encoded network file or NULL if no graph file submitted.
}
\description{
Helper method to encode custom network file
}
