% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{igraph_to_sif}
\alias{igraph_to_sif}
\title{Converts igraph object to sif file.}
\usage{
igraph_to_sif(biological_netwrok, path)
}
\arguments{
\item{path}{Path where the file should be saved.}

\item{biological_network}{Biologial network as igraph object.}
}
\description{
Converts igraph object to sif file.
}
