% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{import_graph}
\alias{import_graph}
\title{Reads graph file and converts it to igraph object}
\usage{
import_graph(file, format, sep)
}
\arguments{
\item{file}{Path to file}

\item{format}{Fomrat of the File}

\item{sep}{You can also use a custom seperator like  TAB or SPACE}
}
\value{
igraph object from the network
}
\description{
Possible formats are c("sif","gml","graphml", "xlsx","custom").
}
\details{
\strong{Important}: For the sif format please also specify if the sep is a TAB or a SPACE.
}
