% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{is_finished}
\alias{is_finished}
\title{Check if asynchronous remote run is finished}
\usage{
is_finished(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
TRUE if results are ready. FALSE if still processing.
}
\description{
Only for remote execution
}
