% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{kpm_options}
\alias{kpm_options}
\title{Set or get options for Rkpm package}
\usage{
kpm_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set argument for the run. To overview all options
use kpm_options(). To get a specific_value from the
arguments use kpm_options()$specific_value.
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{a}}{(\code{numeric};1) The value of a }
 \item{\code{b}}{(\code{numeric};2) The value of b }
}
}

