% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pathway_comparison_plots}
\alias{pathway_comparison_plots}
\title{Two pathway comparison plots.}
\usage{
pathway_comparison_plots(result)
}
\arguments{
\item{result}{Result object from KeyPathwayMineR execution.}
}
\value{
Return two ggplots objects with the comparison plots and
        the data for creating the plots.
}
\description{
The first compares the union network
of each configuration.
The second compares the pathway with the
highest number of nodes of each configuration.
}
