% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{quest_progress_url}
\alias{quest_progress_url}
\title{Get progress url}
\usage{
quest_progress_url(quest_id)
}
\arguments{
\item{quest_id}{}
}
\value{
Returns progress url for the respective run.
}
\description{
Get url to see progress in the browser and see the results
}
