% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{read_graph}
\alias{read_graph}
\title{Reads graph file and converts it to igraph object}
\usage{
read_graph(file, format, sep)
}
\arguments{
\item{file}{Path to file}

\item{format}{Possible formats are c("sif","gml","graphml", "xlsx","custom")}

\item{sep}{You can also use a custom seperator like  TAB or SPACE}
}
\value{
igraph object
}
\description{
Reads graph file and converts it to igraph object
}
