% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{remove_configuration}
\alias{remove_configuration}
\title{Removes configuration from result object}
\usage{
remove_configuration(result_object, configuration_name)
}
\arguments{
\item{result_object}{Result object obtained from a KeyPathwayMineR execution.}

\item{configuration_name}{Configuration name that should be removed.}
}
\description{
Removes configuration from result object
}
