% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{Result-class}
\alias{Result-class}
\title{Result}
\description{
Class for storing the results of a KPM local
or remote execution.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List with the parameters of the current run.}

\item{\code{configurations}}{List with all parameter-exceptions-configurations (K and L) and corresponding pathways.}
}}

