% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{ResultRemote-class}
\alias{ResultRemote-class}
\title{ResultRemote}
\description{
Contains one additional slot which saves
the json as returned by the server.
}
\section{Slots}{

\describe{
\item{\code{json_result}}{json_result returned from the restfulAPI.}
}}

