% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_biogrid}
\alias{retrieve_biogrid}
\title{Retrieves bioGRID network in suitable form for KPM}
\usage{
retrieve_biogrid(organism, IDType)
}
\arguments{
\item{organism}{available organsims are 'arabidopsis', 'c.elegans', 'fruitFly', 'human', 'mouse', 'yeast', 's.pombe'}

\item{IDType}{has to be either 'EntrezId' or 'Official' (=HGNC)}
}
\value{
returns a dataframe which can be used as input in kpm()
}
\description{
Retrieves bioGRID network in suitable form for KPM
}
