% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{save_local_results}
\alias{save_local_results}
\title{Save results of the local run}
\usage{
save_local_results(path_to_results)
}
\arguments{
\item{path_to_results}{The path to the results folder of the local run}
}
\value{
Result object
}
\description{
Method to save the results provided
 from the lccal run.
}
