% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{save_remote_results}
\alias{save_remote_results}
\title{Save results of the remote run}
\usage{
save_remote_results(remote_results)
}
\arguments{
\item{remote_results}{The results returned from the remote run}
}
\value{
Result object
}
\description{
Method to save the results provided
 from the remote run.
}
