% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_singlecell.R
\name{sc_to_indicator_matrix}
\alias{sc_to_indicator_matrix}
\title{Calculates differentially expressed genes from single cell object}
\usage{
sc_to_indicator_matrix(
  sc_obj,
  covarariate,
  referenceGroup,
  saveSummary = TRUE,
  summaryName = NULL,
  designFormula = NULL,
  FCThreshold = 2,
  pvalueThreshold = 0.05,
  nameIndicatorMatrix = NULL
)
}
\arguments{
\item{sc_obj}{single cell object of type seurat, singleCellAssay or singlCellExperiment}

\item{covarariate}{colname of metadata with groups for differential expression}

\item{referenceGroup}{reference/control group}

\item{saveSummary}{if TRUE saves the result of the differential analysis as a .rds file}

\item{summaryName}{name of the rds file if saveSummary=TRUE}

\item{designFormula}{string which will be used as formula in MAST's zlm function}

\item{FCThreshold}{Threshold of log 2 fold change}

\item{pvalueThreshold}{p value cut off after FDR correction for differentially expressed genes}

\item{nameIndicatorMatrix}{indicator matrix which can be used for kpm() execution}
}
\value{
dataframe of of putative differentially expressed genes
}
\description{
Calculates differentially expressed genes between multiple groups with reference to a control group
using a hurdle model tailored to scRNA-seq data
}
\details{
To use this method, MAST has to be installed (see https://github.com/RGLab/MAST/ )
}
\references{
Andrew McDavid, Greg Finak and Masanao Yajima (2017). MAST: Model-based
Analysis of Single Cell Transcriptomics. R package version 1.2.1.
https://github.com/RGLab/MAST/
}
