% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{set_avg_exp}
\alias{set_avg_exp}
\alias{set_avg_exp,Pathway-method}
\title{Set average differential expressed cases per gene}
\usage{
set_avg_exp(pathway, new_avg_exp)

\S4method{set_avg_exp}{Pathway}(pathway, new_avg_exp)
}
\arguments{
\item{pathway}{Pathway for which to set the value.}

\item{new_avg_exp}{New average expression value.}
}
\description{
Set average differential expressed cases per gene
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Sets avg_exp for a specific pathway.
}}

