% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{set_nodes}
\alias{set_nodes}
\alias{set_nodes,Pathway-method}
\title{Set number of nodes given pathway object}
\usage{
set_nodes(pathway, num_nodes)

\S4method{set_nodes}{Pathway}(pathway, num_nodes)
}
\arguments{
\item{pathway}{Pathway for which to set the number of nodes}

\item{num_nodes}{Number of nodes}
}
\description{
Set number of nodes given pathway object
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Sets num_nodes for a specific pathway.
}}

