% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{set_num_edges}
\alias{set_num_edges}
\alias{set_num_edges,Pathway-method}
\title{Set number of edges given pathway object}
\usage{
set_num_edges(pathway, num_edges)

\S4method{set_num_edges}{Pathway}(pathway, num_edges)
}
\arguments{
\item{pathway}{Pathway for which to set the number of edges.}

\item{num_edges}{Number of edges.}
}
\description{
Set number of edges given pathway object
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Sets num_edges for a specific pathway.
}}

