% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{set_pathway}
\alias{set_pathway}
\alias{set_pathway,Result-method}
\title{Set pathway}
\usage{
set_pathway(
  result_object,
  configuration_name,
  pathway_name = "",
  pathway,
  union = FALSE
)

\S4method{set_pathway}{Result}(
  result_object,
  configuration_name,
  pathway_name = "",
  pathway,
  union = FALSE
)
}
\arguments{
\item{union}{Boolean. Whether the pathway is a union network.}
}
\description{
Set pathway
}
\section{Methods (by class)}{
\itemize{
\item \code{Result}: Set pathway in the result object.
}}

