% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful_utils.R
\name{setup.KPM}
\alias{setup.KPM}
\title{Function to set up a JSON object in preparation of the job submission}
\usage{
setup.KPM(
  indicator_matrices,
  algorithm = "Greedy",
  strategy = "GLONE",
  graphID = 1,
  graph.file,
  removeBENs = FALSE,
  range,
  Kmin = 0,
  Lmin = 0,
  Kmax = 0,
  Lmax = 0,
  Kstep = 1,
  Lstep = 1,
  l_same_percentage = FALSE,
  same_percentage = 0,
  attached_to_id,
  computed.pathways = 20,
  with.perturbation = FALSE,
  unmapped_nodes = "Add to negative list",
  linkType = "OR"
)
}
\arguments{
\item{indicator_matrices}{List of indicator matrices.}

\item{algorithm}{A logical scalar. Should missing values (including NaN)}
}
\description{
Function to set up a JSON object in preparation of the job submission
}
