% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{sif_to_igraph}
\alias{sif_to_igraph}
\title{Convert sif file to iGraph object}
\usage{
sif_to_igraph(filepath, sep = sep)
}
\arguments{
\item{filepath}{Filepath to the sif file}

\item{sep}{Either TAB or SPACE}
}
\value{
igraph object from the network
}
\description{
Convert sif file to iGraph object
}
