% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{submit_kpm}
\alias{submit_kpm}
\title{Method for submitting a job to KeyPathwayMinerWeb
Submits can be submitted asynchronously (returns immediately)
or blocked until completed(returns when job is complete).}
\usage{
submit_kpm(kpmSetup)
}
\arguments{
\item{kpmSetup}{kpmSetup}
}
\value{
Results in json format.
}
\description{
Method for submitting a job to KeyPathwayMinerWeb
Submits can be submitted asynchronously (returns immediately)
or blocked until completed(returns when job is complete).
}
