% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_jvm}
\alias{test_jvm}
\title{Checks the version of the JVM.}
\usage{
test_jvm()
}
\value{
TRUE if Java version 1.8 or higher is detected, FALSE otherwise
}
\description{
Checks the version of the JVM.
}
\examples{
test_jvm()
}
