% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matrix.R
\name{to_indicator_matrix}
\alias{to_indicator_matrix}
\title{Convertes numerical matrix to an indicator matrix}
\usage{
to_indicator_matrix(numerical_matrix, operator = "<", threshold = 0.05)
}
\arguments{
\item{numerical_matrix}{Numerical matrix provided as data.frame.
The first row consists of gene_ids and the matrix must not have a header.}

\item{operator}{Operator for converting the matrix. Either "<" or ">".}

\item{threshold}{Threshold to use for converting the matrix.}
}
\value{
Indicator matrix.
}
\description{
Users can select numerical matrix and upload a matrix of numerical values, e.g. p-values or fold changes.
These can then be converted to an indicator matrix on the fly with a user-defined threshold.
}
