% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{to_java_arguments}
\alias{to_java_arguments}
\title{Method which transform kpm_options() to java arguments.}
\usage{
to_java_arguments(indicator_matrices, graph_file)
}
\value{
String [] args array with the arguments
}
\description{
Takes kpm_options and adds arguments to a list
}
