% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_result}
\alias{visualize_result}
\title{Shiny App for KeyPathwayMineR results}
\usage{
visualize_result(result)
}
\arguments{
\item{result}{Result object obtained from a KeyPathwayMiner execution.}
}
\description{
App for visualization and extraction of
KPM results.
}
