% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{withTryCatch}
\alias{withTryCatch}
\title{Helper method for error handling}
\usage{
withTryCatch(surroundedFunc)
}
\arguments{
\item{surroundedFunc}{Function in which the TryCatch error handling should be applied.}
}
\description{
Helper method for error handling
}
