% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{call_kpm_remote}
\alias{call_kpm_remote}
\title{Method used to create a job submission}
\usage{
call_kpm_remote(indicator_matrices)
}
\arguments{
\item{indicator_matrices}{List. Indicator matrices that will be used.}

\item{attached_to_id}{String. The quest id this job has been attached to.}

\item{async}{Boolean. Submit a new job asynchronously?}

\item{...}{Options. Run parameters.}
}
\value{
Results object with runId, resultGraphs, comment, succes,resultUrl
 and questId.
}
\description{
Method used to create a job submission
}
\examples{
call_kpm_remote
        async=FALSE, Lmin=8, Kmin=1, strategy="INES",
        removeBENs=TRUE, graphID=I2D.id, graph.file=NULL,
        range=FALSE, linkType="OR")
}
