% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{call_kpm_remote}
\alias{call_kpm_remote}
\title{Method for running KeyPathwayMineR remotely}
\usage{
call_kpm_remote(matrices, graph_file = NULL)
}
\arguments{
\item{matrices}{List. Indicator matrices that will be used.}

\item{graph_file}{String. Path to graph_file.}
}
\value{
Results object with runId, resultGraphs, comment, succes, resultUrl
 and questId.
}
\description{
Method used to create a job submission.
}
