% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpm.R
\name{check_files}
\alias{check_files}
\title{Function which checks and processes files}
\usage{
check_files(indicator_matrices, graph_file)
}
\arguments{
\item{indicator_matrices}{Filepath, data.frame or as a list of both. Indicator matrices.}

\item{graph_file}{Filepath to the graph file in sif format.}
}
\value{
Returns the data prepared for the respective run (local or remote)
}
\description{
Goes through all matrices and the graph file and checks
if the input data is correct and exists. It then prepares
the data for local or remote use.
}
