% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{dataset_list}
\alias{dataset_list}
\title{Encodes a list of datasets}
\usage{
dataset_list(indicator_matrices)
}
\arguments{
\item{indicator_matrices}{List of indicator matrices to be encoded}
}
\value{
Encoded list of datasets with their respective questId.
}
\description{
Helper method to encode a list of datasets (indicator matrices)
 for the job submission.
}
