% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{export_graph}
\alias{export_graph}
\alias{export_graph,Pathway-method}
\title{Export a graph}
\usage{
export_graph(pathway_object, file, sep = "\\t", format = "")

\S4method{export_graph}{Pathway}(pathway_object, file, sep = "\\t", format = "")
}
\arguments{
\item{pathway_object}{pathway object of a configuration.}

\item{file}{path where the file is saved}

\item{sep}{seperator used to seperate the graph file.}

\item{format}{format of the file. If specified sep slot is not used.}
}
\description{
Export a graph in sif format, tab seperated or custom format.
}
\section{Methods (by class)}{
\itemize{
\item \code{Pathway}: Use write.table to export a graph.
}}

