% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_pathway}
\alias{get_pathway}
\alias{get_pathway,Configuration-method}
\title{Get specific pathway of a configuration}
\usage{
get_pathway(configuration, pathway_name = "", union = FALSE)

\S4method{get_pathway}{Configuration}(configuration, pathway_name = "", union = FALSE)
}
\arguments{
\item{configuration}{Specific configuration for which to get the pathways.}

\item{pathway_name}{Name of the pathway to be extracted.}

\item{union}{Boolean. Whether the pathway is a union network.}
}
\value{
Pathway from given configuration with specific pathway name.
}
\description{
Get specific pathway of a configuration
}
\section{Methods (by class)}{
\itemize{
\item \code{Configuration}: Return pathway with specific pathway name.
}}

