% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{get_pathways}
\alias{get_pathways}
\alias{get_pathways,Result-method}
\title{Get all pathways for a specific configuration}
\usage{
get_pathways(result_object, configuration)

\S4method{get_pathways}{Result}(result_object, configuration)
}
\arguments{
\item{result_object}{Result of the current run.}

\item{configuration}{Specific configuration for which to get the pathways.}
}
\value{
All pathways of a specific configuration.
}
\description{
Get all pathways for a specific configuration
}
\section{Methods (by class)}{
\itemize{
\item \code{Result}: Returns a list of pathways for a specific coniguration
}}

