% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{get_result_url}
\alias{get_result_url}
\title{Returns result url}
\usage{
get_result_url(result_object)
}
\description{
Only for remote execution
}
\details{
@param result_object Result of the current run.

 @return Url to the results on the KPM web server.
 @export
}
