% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{get_results}
\alias{get_results}
\title{Fetches result for remote run.}
\usage{
get_results(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
Remote result object.
}
\description{
Only for remote execution.
The job must be completed to get the results.
Completion can be checked with is_finished(result_object).
}
