% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_utils.R
\name{get_results}
\alias{get_results}
\title{Once the run is complete, we can obtain the results}
\usage{
get_results(quest_id)
}
\arguments{
\item{quest_id}{Jobs respective quest_id.}
}
\value{
If run was successful return result in json
format otherwise null.
}
\description{
Once the run is complete, we can obtain the results
}
