% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{is_finished}
\alias{is_finished}
\alias{is_finished,ResultRemote-method}
\title{Check if asynchronous remote run is finished}
\usage{
is_finished(result_object)

\S4method{is_finished}{ResultRemote}(result_object)
}
\arguments{
\item{result_object}{Result of the current run.}
}
\value{
TRUE if results are ready. FASE if still processing .
}
\description{
Only for remote execution
}
\section{Methods (by class)}{
\itemize{
\item \code{ResultRemote}: returns status of remote run
}}

