% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpm.R
\name{kpm}
\alias{kpm}
\title{Main method used for running KPM}
\usage{
kpm(indicator_matrices, graph_file = NULL)
}
\arguments{
\item{indicator_matrices}{List of paths to the indicator matrices.}

\item{graph_file}{Path of the graph file. NULL if you want to use
a graph from the website (only for remote runs).
Use getNetworks('https://exbio.wzw.tum.de/keypathwayminer/')
to see all networks.}
}
\description{
Runs KeyPathwayMiner localy via the standalone
or remotely via the RESTful API of the KeyPathwayMiner
website. Input parameters are indicator matrices and a graph_file.
You can view or change the run parameters through the
kpm_options() function.
}
