% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{kpm_options}
\alias{kpm_options}
\title{Set or get options for my package}
\usage{
kpm_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{a}}{(\code{numeric};1) The value of a }
 \item{\code{b}}{(\code{numeric};2) The value of b }
}
}

