% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{Pathway-class}
\alias{Pathway-class}
\title{A pathway of a specific configuration.}
\description{
A pathway of a specific configuration.
}
\section{Slots}{

\describe{
\item{\code{edges}}{Edges of the pathway.}

\item{\code{nodes}}{Nodes of the pathway.}

\item{\code{exception_nodes}}{Exception nodes in the pathway. Indicator list with 1 indicating a significant node and 0 an exception node. Should be in the same order as the entities in the nodes slot.}

\item{\code{num_edges}}{Number of edges.}

\item{\code{num_nodes}}{Number of nodes.}

\item{\code{exception}}{Indicator list of  Number of exception nodes in the pathway.}

\item{\code{avg_exp}}{Average differentially expressed cases per gene.}
}}

