% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{pathway_statistics}
\alias{pathway_statistics}
\title{Computes pathway statistics for all pathways in a result object.}
\usage{
pathway_statistics(indicator_matrix, result)
}
\arguments{
\item{indicator_matrix}{Indicator matrix for the computation of the pathway statistics.}

\item{result}{Result object obtained from a KeyPathwayMiner execution.}
}
\description{
Computes pathway statistics for all pathways in a result object.
}
