% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{setup_kpm}
\alias{setup_kpm}
\title{Set up a JSON object in preparation of the job submission}
\usage{
setup_kpm(indicator_matrices, graph_file)
}
\value{
A list with settings, datasets and graph in JSON format.
}
\description{
Set up a JSON object in preparation of the job submission
}
