% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_jvm}
\alias{test_jvm}
\title{Fcuntion for testing if rJava works}
\usage{
test_jvm()
}
\value{
TRUE if the correct Java version is detected, FALSE otherwise
}
\description{
Fcuntion for testing if rJava works
}
\examples{
test_jvm()
}
